refactor:  Separate HttpSocket class in hpp and cpp files
